src/bookkeeper_checklist_project/task/serializers/task_item.py
# -*- coding: utf-8 -*-#
from rest_framework import serializers

from core.constants import EXCLUDED_FIELDS
from task.models import TaskItem


class TaskItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = TaskItem
        exclude = EXCLUDED_FIELDS
