requirements/dev.txt
-r _base.txt
autopep8
bandit
black
black[d]
bpython
django-debug-toolbar
django-extensions
django-stubs
django-stubs-ext
djangorestframework-stubs
djangorestframework-stubs[compatible-mypy]
django-request-viewer
djlint
docformatter
faker
flake8
inquirer
isort
mypy
mypy-extensions
pep8
phone-gen
pip-chill
pipdeptree
prettyprinter
prettytable
pycodestyle
pydocstyle
pygraphviz
pylama
pylint
pylint-django
PyPika
pytest
pytest-django
remote-pdb
rich
Selenium
termcolor
yapf
